+++ app/Http/Controllers/admin/RoomsTypeController.php
@@ -3,6 +3,8 @@
 namespace App\Http\Controllers\admin;
 
 use App\Http\Controllers\Controller;
+use App\Http\Requests\typeRoom\TypeRoomStoreRequest;
+use App\Models\rooms;
 use Illuminate\Http\Request;
 use App\Models\type_rooms;
 
@@ -12,7 +14,7 @@ class RoomsTypeController extends Controller
      * Display a listing of the resource.
      */
     public function index()
-    {  
+    {
         $listTypeRooms = type_rooms::all();
         return view('admin.rooms_type.index', compact('listTypeRooms'));
     }
@@ -28,11 +30,10 @@ class RoomsTypeController extends Controller
     /**
      * Store a newly created resource in storage.
      */
-    public function store(Request $request)
+    public function store(TypeRoomStoreRequest $request)
     {
         type_rooms::create($request->all());
-        return redirect()->route('rooms_type.index');
-
+        return redirect()->route('rooms_type.index')->with('success','Thêm mới thành công');
     }
 
     /**
@@ -48,18 +49,18 @@ class RoomsTypeController extends Controller
      */
     public function edit(type_rooms $rooms_type)
     {
-        return view('admin.rooms_type.edit',compact('rooms_type'));
+        return view('admin.rooms_type.edit', compact('rooms_type'));
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, type_rooms $rooms_type)
+    public function update(TypeRoomStoreRequest $request, type_rooms $rooms_type)
     {
         try {
             //code...
             $rooms_type->update($request->all());
-            return redirect()->route('rooms_type.index')->with('success','Update successful');
+            return redirect()->route('rooms_type.index')->with('success', 'Cập nhật thành công');
         } catch (\Throwable $th) {
             dd($th);
         }
@@ -70,19 +71,25 @@ class RoomsTypeController extends Controller
      */
     public function destroy(string $id)
     {
-        type_rooms::where('id',$id)->delete();
-        return redirect()->route('rooms_type.index');
+        if (rooms::where('type_id', $id) == '') {
+            type_rooms::where('id', $id)->delete();
+            return redirect()->route('rooms_type.index')->with('success','Xoá thành công');
+        }
+        return redirect()->back()->with('error','Hãy xoá phòng trước khi xoá hạng phòng');
     }
-    public function trash(){
+    public function trash()
+    {
         $deleted_at = type_rooms::onlyTrashed()->get();
-        return view('admin.rooms_type.trash',compact('deleted_at'));
+        return view('admin.rooms_type.trash', compact('deleted_at'));
     }
-    public function restore($id) {
-        type_rooms::withTrashed()->where('id',$id)->restore();
+    public function restore($id)
+    {
+        type_rooms::withTrashed()->where('id', $id)->restore();
         return redirect()->route('rooms_type.index');
     }
-    public function forceDeleted($id) {
-        type_rooms::withTrashed()->where('id',$id)->forceDelete();
+    public function forceDeleted($id)
+    {
+        type_rooms::withTrashed()->where('id', $id)->forceDelete();
         return redirect()->route('rooms_type.trash');
     }
 }
